moved the duplicate code to common file
